refact: community comment 페이지 훅 분리
